+++ examples/core/rtmp_publish/src/main.rs
@@ -71,6 +71,7 @@ mod example {
                 codec: VideoCodec::OpenH264,
                 width: video_options.width,
                 height: video_options.height,
+                pixel_format: ffmpeg::format::Pixel::YUV420P,
                 frame_rate: video_options.frame_rate,
                 bit_rate: 1_500_000,
                 gop_size: GOP_SIZE,
@@ -138,7 +139,7 @@ mod example {
             let pipeline = pipeline.clone();
             thread::spawn(move || {
                 thread::sleep(Duration::from_secs(seconds));
-                pipeline.stop();
+                pipeline.finish();
             });
         }