+++ examples/core/video_compositor/src/main.rs
@@ -137,6 +137,7 @@ mod example {
                 codec: VideoCodec::OpenH264,
                 width: output_width,
                 height: output_height,
+                pixel_format: ffmpeg::format::Pixel::YUV420P,
                 frame_rate,
                 bit_rate: 2_000_000,
                 gop_size: 60,
@@ -194,7 +195,7 @@ mod example {
 
         background_pipeline.stop();
         foreground_pipeline.stop();
-        output_pipeline.stop();
+        output_pipeline.finish();
 
         for pipeline in [&background_pipeline, &foreground_pipeline, &output_pipeline] {
             for event in pipeline.bus().iter() {