+++ examples/render/gpu_video_compositor/src/main.rs
@@ -182,6 +182,7 @@ mod windows_example {
                 codec: VideoCodec::OpenH264,
                 width: output_width,
                 height: output_height,
+                pixel_format: ffmpeg::format::Pixel::YUV420P,
                 frame_rate,
                 bit_rate: 2_000_000,
                 gop_size: 60,
@@ -268,7 +269,7 @@ mod windows_example {
 
         background_pipeline.stop();
         foreground_pipeline.stop();
-        output_pipeline.stop();
+        output_pipeline.finish();
 
         for pipeline in [&background_pipeline, &foreground_pipeline, &output_pipeline] {
             for event in pipeline.bus().iter() {
@@ -436,6 +437,7 @@ mod linux_example {
                 codec: VideoCodec::OpenH264,
                 width: output_width,
                 height: output_height,
+                pixel_format: ffmpeg::format::Pixel::YUV420P,
                 time_base,
                 frame_rate,
                 bit_rate: 2_000_000,
@@ -525,7 +527,7 @@ mod linux_example {
 
         background_pipeline.stop();
         foreground_pipeline.stop();
-        output_pipeline.stop();
+        output_pipeline.finish();
 
         for pipeline in [&background_pipeline, &foreground_pipeline, &output_pipeline] {
             for event in pipeline.bus().iter() {